requirements.txt
appdirs==1.4.0
facebook-sdk==2.0.0
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
